+++ src/include/smash/listmodus.h
@@ -62,14 +62,16 @@ class ListModus : public ModusDefault {
    *
    * \param[in] modus_config The configuration object that sets all
    *                         initial conditions of the experiment.
-   * \param[in] parameters Unused, but necessary because of templated
-   *                       initialization
+   * \param[in] parameters Necessary because of templated usage in Experiment.
    */
   explicit ListModus(Configuration modus_config,
                      const ExperimentParameters &parameters);
 
-  /// Construct an empty list. Useful for convenient JetScape connection.
-  ListModus() : shift_id_(0) {}
+  /**
+   * Construct an empty list. This is needed for children construction but it is
+   * offered as public instead of protected as it is also useful for JetScape.
+   */
+  ListModus() = default;
 
   /**
    * Generates initial state of the particles in the system according to a list.
@@ -145,34 +147,17 @@ class ListModus : public ModusDefault {
   /// \return whether the modus is list modus (which is, yes, trivially true)
   bool is_list() const { return true; }
 
-  /// set the file id when ListBoxModus is used
-  void set_file_id(const double file_id_inh) { file_id_ = file_id_inh; }
-
-  /// set the particle_list_directory when ListBoxModus is used
-  void set_particle_list_file_directory(
-      std::string particle_list_file_directory_inh) {
-    particle_list_file_directory_ = particle_list_file_directory_inh;
-  }
-
-  /// set the particle_list_prefix when ListBoxModus is used
-  void set_particle_list_file_prefix(
-      std::string particle_list_file_prefix_inh) {
-    particle_list_file_prefix_ = particle_list_file_prefix_inh;
-  }
-
-  /// set the event_id when ListBoxModus is used
-  void set_event_id(int event_id_inh) { event_id_ = event_id_inh; }
-
  protected:
   /// Starting time for the List; changed to the earliest formation time
   double start_time_ = 0.;
 
  private:
-  /** Check if the file given by filepath has events left after streampos
+  /**
+   * Check if the file given by filepath has events left after streampos
    * last_position
    *
    * \param[in] filepath Path to file to be checked.
-   * \param[in] last_position Streamposition in file after which check is
+   * \param[in] last_position Stream position in file after which check is
    * performed
    * \return True if there is at least one event left, false otherwise
    * \throws runtime_error If file could not be read for whatever reason.
@@ -180,17 +165,22 @@ class ListModus : public ModusDefault {
   bool file_has_events_(std::filesystem::path filepath,
                         std::streampos last_position);
 
-  /** Return the absolute file path based on given integer. The filename
-   * is assumed to have the form (particle_list_prefix)_(file_id)
+  /**
+   * Return the absolute path of the data file. If an integer is passed, the
+   * filename is constructed using \c particle_list_filename_or_prefix_
+   * concatenated with the given number, otherwise the file prefix is understood
+   * to be the full filename. The file is expected to be in \c
+   * particle_list_file_directory_ folder.
    *
-   * \param[in] file_id integer of wanted file
+   * \param[in] file_id An \c std::optional integer
    * \return Absolute file path to file
-   * \throws
-   * runtime_error if file does not exist.
+   *
+   * \throws runtime_error if file does not exist.
    */
-  std::filesystem::path file_path_(const int file_id);
+  std::filesystem::path file_path_(std::optional<int> file_id);
 
-  /**  Read the next event. Either from the current file if it has more events
+  /**
+   * Read the next event. Either from the current file if it has more events
    * or from the next file (with file_id += 1)
    *
    * \returns
@@ -202,28 +192,25 @@ class ListModus : public ModusDefault {
   /// File directory of the particle list
   std::string particle_list_file_directory_;
 
-  /// File prefix of the particle list
-  std::string particle_list_file_prefix_;
+  /**
+   * Prefix of the file(s) containing the particle list. If the user want to
+   * use a single file without numbering, this will contain the full filename.
+   */
+  std::string particle_list_filename_or_prefix_;
 
-  /// File name of current file
-  std::string current_particle_list_file_;
+  /// The id of the current file
+  std::optional<int> file_id_;
 
-  /// shift_id is the start number of file_id_
-  const int shift_id_;
-
-  /// event_id_ = the unique id of the current event
+  /// The unique id of the current event
   int event_id_;
 
-  /// file_id_ is the id of the current file
-  int file_id_;
+  /// Last read position in current file
+  std::streampos last_read_position_ = 0;
 
   /// Auxiliary flag to warn about mass-discrepancies only once per instance
   bool warn_about_mass_discrepancy_ = true;
   /// Auxiliary flag to warn about off-shell particles only once per instance
   bool warn_about_off_shell_particles_ = true;
-
-  /// last read position in current file
-  std::streampos last_read_position_;
 
   /**\ingroup logging
    * Writes the initial state for the List to the output stream.
@@ -312,23 +299,8 @@ class ListBoxModus : public ListModus {
   }
 
  private:
-  /// shift_id is the start number of file_id_
-  const int shift_id_;
-
   /// Length of the cube's edge in fm
   const double length_;
-
-  /// file_id_ is the id of the current file
-  int file_id_;
-
-  /// event_id_ = the unique id of the current event
-  int event_id_;
-
-  /// File directory of the particle list
-  std::string particle_list_file_directory_;
-
-  /// File prefix of the particle list
-  std::string particle_list_file_prefix_;
 };
 
 }  // namespace smash