+++ src/tests/listmodus.cc
@@ -105,7 +105,27 @@ static ListModus create_list_modus_for_test() {
     List:
       File_Directory: ToBeSet
       File_Prefix: event
-      Shift_Id: 0
+    )"};
+  config.set_value({"List", "File_Directory"}, testoutputpath.string());
+  return ListModus(std::move(config), parameters);
+}
+
+static ListBoxModus create_list_box_modus_for_test() {
+  Configuration config{R"(
+    ListBox:
+      File_Directory: ToBeSet
+      File_Prefix: event
+      Length: 3
+    )"};
+  config.set_value({"ListBox", "File_Directory"}, testoutputpath.string());
+  return ListBoxModus(std::move(config), parameters);
+}
+
+static ListModus create_list_modus_with_single_file_for_test() {
+  Configuration config{R"(
+    List:
+      File_Directory: ToBeSet
+      Filename: event0
     )"};
   config.set_value({"List", "File_Directory"}, testoutputpath.string());
   return ListModus(std::move(config), parameters);
@@ -169,6 +189,16 @@ TEST(list_from_non_oscar_output) {
     COMPARE_ABSOLUTE_ERROR(a.formation_time(), b.formation_time(), accuracy);
     COMPARE(a.pdgcode(), b.pdgcode());
   }
+}
+
+TEST(listbox_creation_from_non_oscar_output) {
+  std::vector<ParticleList> init_particles;
+  create_non_oscar_particlefile(0, init_particles);
+  ListBoxModus list_box_modus = create_list_box_modus_for_test();
+
+  // Read the file with list modus
+  Particles particles_read;
+  list_box_modus.initial_conditions(&particles_read, parameters);
 }
 
 TEST(multiple_file_non_oscar_output) {
@@ -384,7 +414,7 @@ TEST(multiple_events_in_file) {
   std::vector<ParticleList> init_particles;
   create_particlefile(out_par, 0, init_particles, particles_per_event,
                       max_events);
-  ListModus list_modus = create_list_modus_for_test();
+  ListModus list_modus = create_list_modus_with_single_file_for_test();
 
   for (int cur_event = 0; cur_event < max_events; cur_event++) {
     // Read the file with list modus