+++ CHANGELOG.md
@@ -22,19 +22,23 @@ Also possible, but for this project less relevant, is `Deprecated` for soon-to-b
 
 ### Added
 * Option `Decay_Initial_Particles` which blocks the decay of resonances created at initialization
+* The `List` and `ListBox` modi can now be used specifying a single particles input file using the `Filename` key
 
 ### Fixed
 * Fix bug in the time step calculation of SMASH output in events successive to the first one when the `End_Time` is not a multiple of the `Output_Interval`
 * SMASH now aborts with a user-friendly message if any input key among `End_Time`, `Delta_Time` or `Output_Interval` is zero
 * Fix bug in thermodynamics lattice ASCII output when both ASCII and binary formats are requested
-* Fix bug about SMASH getting stuck with some particular values of `End_Time` and `Delta_Time`.
+* Fix bug about SMASH getting stuck with some particular values of `End_Time` and `Delta_Time`
 
 ### Changed
 * Upgraded to Pythia 8.310
 * Upgraded to Cuba 4.2.2
 * SMASH installation procedure was improved and allows now to use SMASH as library relying on installation folder only
-* Make `Clock` class comparison operators refer to internally represented time and not to `current_time()` return value.
+* Make `Clock` class comparison operators refer to internally represented time and not to `current_time()` return value
 * Add strangeness to all extended output formats in the particle line
+
+### Removed
+* Some public setters of the `ListModus` class have been removed
 
 
 ## SMASH-3.0